+++ NEWS.md
@@ -4,7 +4,11 @@ Pentobi Release Notes
 Version 21.0-dev (master)
 -------------------------
 
-*
+* New help viewer implementation, which includes DocBook files in the
+  resources. This avoids using an external browser, which did not work
+  with all installation directories on Ubuntu 21.10 because Firefox
+  and Chromium have restricted access to local files, or using WebView,
+  which is not supported on all Linux architectures or in Qt 6.0/6.1.
 
 Version 20.1-dev (v20-fixes)
 ----------------------------